refactored words into seperate components
