--- src/components/ItemTypes.js
@@ -1,4 +1,0 @@
-export default {
-  SECONDARY_WORD: 'secondary',
-  PRIMARY_WORD: 'primary'
-};
