+++ src/components/Word/Types.js
@@ -1,0 +1,2 @@
+export const SECONDARY_WORD = 'secondary';
+export const PRIMARY_WORD = 'primary';
