+++ src/components/Word/index.js
@@ -1,66 +1,62 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Word from './Word';
-import { DragSource } from 'react-dnd';
-import ItemTypes from '../ItemTypes';
+import WordOccurrence from './WordOccurrence';
 
-/**
- * Renders a draggable word.
- *
- * @property {string} word - the represented word
- * @property {int} occurrence
- * @property {int} occurrences
- */
-class DraggableWord extends React.Component {
-  render() {
-    const {connectDragSource, word, occurrence, occurrences, isDragging} = this.props;
-    const opacity = isDragging ? 0.4 : 1;
-
-    return connectDragSource(
-      <div style={{flex:1}}>
-        <Word word={word}
-              style={{opacity}}
-              occurrence={occurrence}
-              occurrences={occurrences}/>
-      </div>
-    );
-  }
-}
-
-DraggableWord.propTypes = {
-  word: PropTypes.string.isRequired,
-  occurrence: PropTypes.number.isRequired,
-  occurrences: PropTypes.number.isRequired,
-  connectDragSource: PropTypes.func.isRequired,
-  isDragging: PropTypes.bool.isRequired,
-  alignmentIndex: PropTypes.number
-};
-
-DraggableWord.defaultProps = {
-  alignmentIndex: undefined // just to be explicit
-};
-
-/**
- * Handles drag events on the word
- */
-const dragHandler = {
-  beginDrag(props) {
-    // Return the data describing the dragged item
-    return {
-      word: props.word,
-      occurrence: props.occurrence,
-      occurrences: props.occurrences,
-      alignmentIndex: props.alignmentIndex,
-      type: ItemTypes.SECONDARY_WORD
-    };
+const styles = {
+  root: {
+    borderLeft: '5px solid #44C6FF',
+    padding: '10px',
+    backgroundColor: '#FFFFFF',
+    boxShadow: "0 1px 4px rgba(0, 0, 0, 0.3), 0 0 40px rgba(0, 0, 0, 0.1) inset",
+    cursor: 'move',
+    display: 'flex',
+    flexDirection: 'row'
+  },
+  word: {
+    cursor: 'pointer'
   }
 };
 
-export default DragSource(
-  ItemTypes.SECONDARY_WORD,
-  dragHandler,
-  (connect, monitor) => ({
-    connectDragSource: connect.dragSource(),
-    isDragging: monitor.isDragging()
-  })
-)(DraggableWord);
+/**
+ * Renders a single word
+ * @param {string} word - the represented word
+ * @param {int} occurrence
+ * @param {int} occurrences
+ * @param {object} [style] - styles passed  through to the component
+ * @param {func} [onClick] - callback when the word is clicked
+ * @constructor
+ */
+const Word = ({word, occurrence, occurrences, style, onClick}) => {
+  const rootStyles = {...styles.root, ...style};
+  let wordStyles = null;
+  if(typeof onClick === 'function') {
+    wordStyles = styles.word;
+  }
+  return (
+    <div style={rootStyles}>
+      <span style={{flex: 1}}>
+        <span onClick={onClick} style={wordStyles}>
+          {word}
+        </span>
+      </span>
+      <WordOccurrence occurrence={occurrence}
+                      occurrences={occurrences}/>
+    </div>
+  );
+};
+
+Word.propTypes = {
+  onClick: PropTypes.func,
+  style: PropTypes.object,
+  occurrence: PropTypes.number,
+  occurrences: PropTypes.number,
+  word: PropTypes.string.isRequired
+};
+
+Word.defaultProps = {
+  style: {},
+  occurrence: 1,
+  occurrences: 1
+};
+
+export default Word;
